grunt: target to allow build dir customization
This patch adds a target, `exec:build_includes` to Saucer's Gruntfile. It
also adds a `build_includes` property to the top-level `paths` dictionary that
the user can modify in order copy specific files or directories into
paths.build_dir as part of the build process.

For example:

  paths =

    ...

    build_includes:
      '/'           : [ '<src-path-1>', '<src-path-2>' ]
      '<dest-path>' : [ '<src-path-3>', '<src-path-4>/' ]

The above will copy <src-path-1> and <src-path-2> into build;
<src-path-3> into build/<dest-path>; and the contents of <src-path-4>
into build/<dest-path>.

Note that, if a source path ends in a '/' and points to a directory, the
contents of the directory are copied, rather than the directory itself.